Improve the rest api a bit
This commit adds some more methods to the api to add additional query
support. It adds other top level queries for tests and test runs it
then adds some routes for filtered tests and test_runs by a specific
run. It also adds a query to get a list of runs for a specific
build_name (aka a job name).

Change-Id: Ib0376f39819676b3eb33a2693fad9ddf85aaa509