Add shim for Highcharts.
As Highcharts does not support module loaders, remove its npm
dependency and add it manually as part of a new "vendor-js" build
step. Libraries in `app/vendor-js/` will be concatenated into a final
"vendor.js" file which is included manually into the index page. The
Highcharts module is then exposed via a shim.

Change-Id: I0d3b1d938b85ffe5f54fca7467a55a08507bd155