Add unit tests for rest api and fix docs
This commit adss the basic framework for unit tests of the rest api
and 2, yes 2, unit tests to cover the /runs and /tests endpoints.
This will be expanded in the future, but this is to just give us a
starting point as we expand the api.

Additionally, to unblock the repo this adds missing tox envs which
are needed to pass the docs job. It also takes care of the other
missing pieces to pass a build sphinx run.

Change-Id: Id46476285612789e3b7817eacc45125cf6c3a548