Add charts from mockup to home view.
This adds a new 'data' directory with some sample JSON, as well as
angular directives for the two chart types. Highcharts is also
replaced with Highstock to get access to the stock chart. The
home controller fetches the data file, parses it, and passes it along
to the view for display.

Change-Id: Ia3a9a946c1f6cac2c651c2161199a063edd30326
Depends-On: I0d3b1d938b85ffe5f54fca7467a55a08507bd155