Add rest api method for first page view
This commit adds a new rest endpoint /runs/group_by/metadata_key to
return a json dictionary of the runs stats for all the runs in a given
time resolution grouped by the metadata_key provided in the url.

This will be used for the first page view to return a json dict that
is grouped by project by making a call like: GET /runs/group_by/project
which will return all the runs grouped by project.

Change-Id: I73a9a040a8950b72b399d1866246422b08f2e60d