Add REST API endpoint for second page view
This commit adds a rest api endpoint for the second page view. It
takes in a key value pair from the url and returns a time series dict
of all the runs which had that key value pair in it's respective run
metadata. The datetime resolution, start date and stop date are all
adjustable with url parameters.

The second page view will use this with the key being project and the
value being whatever project the page is being generated for.

Co-Authored-By: Glauco Oliveira <gvinici@thoughtworks.com>
Co-Authored-By: Moises Trovó <mtrovo@thoughtworks.com>

Change-Id: I7837073c9029014e03b2faca642f77f997ebdf82