Add REST API endpoint for the third page view
This commit adds a rest api endpoint for the third page view. It takes
in a build_name from the url and returns a time series dict of all
the test runs which ran as part of a run which were that build_name.
The start date, stop date, and resolution of the time series is
adjustable with url params.

The third page will use this with the key being a project and the
value being whatever project the page is being generated for.

Change-Id: Ic3a74954015dd6b3bcc8eab4c8854e60906e7a85