Add second page (project)
This commit adds the second page to the app which shows detailed job
metrics from a specific project. To access this page user must click on
project name on home page.

For now the page is not acessing api method to get real data, but
a sample json created based on 'project_data.json' in mockups folder.

Change-Id: Ia257bb20f4a73a6e31ea98e1258dbe5fdd42c03d
Co-Authored-By: Glauco Oliveira <gvinici@thoughtworks.com>