Use live data on dashboard homepage
This patch adjusts the home controller and gauge chart to request
and display live data from the REST API, including modifications to
data processing to handle adjustments to the incoming JSON structure.
This patch assumes that a REST server is running at:

http://localhost:5000/

Which is only useful for local dev envs using a local api server. A
later patch will make this configurable so it can be deployed for
real.

Co-Authored-By: Matthew Treinish <mtreinish@kortar.org>
Change-Id: Ic2db942d4905403b6c4b89a7d56ec7794fa106c4