Update flake8 path in tox.ini
The tox.ini pep8 command was running flake8 on the whole repo, however
when also testing the frontend side of the project some of the js
dependencies pull down python code. We don't want to run flake8 on
external code. This commit adjusts the command to only run flake8
on the openstack_health dir which will limit it to our code.

Change-Id: I003e1b87c4990fd09596a222cb506b113eb47abb