Add build failure rate chart on homepage
This change includes a new build failure rate chart on dashboard
homepage. This chart will show the normalized amount of failures when
compared to the amount of total builds.

We believe that this will help us identifying failure trends
independently on the amount of builds running at a given moment.

Change-Id: Ic723a2159e55d61e4648df5e0b01bf37bdf3bdad
Co-Authored-By: Caio Carrara <ccarrara@thoughtworks.com>