Add a rest api to get a test_runs time series by test
This commit adds a new REST API method to get a time series of
test_runs given the test_id for a test. The intent here is to enable
a per test view with a graph of success and failures over time, as
well as a test run_time graphed over time.

Change-Id: Id7fe36c3e1ca069d942fe246d688648f719d3168