Fix project name encoded in project controller
As the project name is passed to the project controller as a part of URL
and URL can be encoded, this commit change the controller to decode the
project name parameter with decodeURIComponent() before anything. It is
necessary both to exibit the name properly and pass it to the server
side api.

Change-Id: I03472c54259daf5c6bcd85dbdd3b59545e391cae