Include d3 and nvd3 as static dependencies.
As the current latest release of nvd3 has a number of major bugs,
we need to use the latest available git release. However, depending
on libraries from `https://` Git repositories is broken on older
versions of npm.

As a workaround, this patch adds nvd3 (and consequently, d3 itself)
as static dependencies using the `vendor-js` mechanism and keeps
the code for each directly within our source tree.

Change-Id: I4fe83ca7eee21dcd1e6d2692f56592e6cb4ea74e