Fix wsgi_script entry point
Previously the wsgi_script openstack-health entry point was pointing
directly to flask app object in the api module. However the script
constructed by pbr attempts to call app and set that as the
application for mod_wsgi. However this is the incorrect usage for a
flask app, and will fail because the app object isn't expected to be
called by itself without any args) Also, according to the flask docs
[1] you just need to set the flask app as the application var in the
to run with mod_wsgi. This commit fixes the wsgi_script entrypoint by
creating a wrapper function which can be called by the generated
output of wsgi scripts from pbr and will return the flask app. This
function is then used as the wsgi_script entrypoint instead of the app
itself. Additionally, this also gives a good injection point if there
are any future initialization or configuration steps that need to be
taken when running.

[1] http://flask.pocoo.org/docs/0.10/deploying/mod_wsgi/#creating-a-wsgi-file

Change-Id: I6e49b4dbbe58e054a28412207b99eb2982a45fbe