Remove mockups for implemented pages
This commit removes the mockups, sample data files, and other js files
related to the mockups which served as guides to getting things up. We
already have working code landed for the first 2 pages and the mockups
don't really reflect what things actually look like any more. So there
really isn't a reason to keep them around anymore. We should also
remember to remove the mockups on the patches which implement the
corresponding page for real.

Change-Id: Ice14238928bcb534d9870ad51eb9585490bfd705