Use PhantomJS and enforce singleRun for tests.
Karma previously started a continuous server to allow for multiple
test runs, which would cause `npm test` to hang when tests finish.
This enables the config option `singleRun` so it quits after tests
have finished running.

Additionally, this removes Chrome and the Chrome launcher so tests
can be run properly in a headless environment.

Change-Id: Ib493d9cac3be5a83aa342b0f2a142f82a2757ffb