Fix date format passed to dbapi
There was some rest endpoint method which wasn't parsing start and stop
date format before calling dbapi method. It was causing warning because
these fields were not in same format as run_at db field. This change
implement the _parse_datetimes call properly where was missing.

It was necessary update the timestamp variable on tests because the
dbapi call is made regardless minutes and seconds.

Change-Id: I3bb4cf4d1efebad0b81926a400dc91c8b647b4fd