Make connection pool settings configurable
When running the api server in production we're seeing a dropped
connection causing 500 responses and the server never recovering
from this. This commit attempts to address this by making the
connection pooling settings configurable and setting better defaults
which will result in not dropping all requests after 1 db connection
timeout.

Change-Id: I1be350794651eb8ec544c0626bb9d516d7efef1b