Add unit tests around HomeController
As part of building a reliable UI, I am including tests around our
core components.

There is a minor change made to HomeController: instead of getting the
current date as part of the loading process, the date is injected once
the controller is loaded. That helps us testing the controller behavior.

Change-Id: Id19bc9e212329990a415077f53924321ce737bed