Stop trying to set defaults with the prescribed method
The config parser docs say you should set default values for missing
config options by passing in a dict to the parser init which will be
interpolated when a non-existent option is used. However, in practice
this either completely didn't work in the case of my local testing, or
in the case of the production server it tried to use it but couldn't
deal with the int values. This commit just drops the use of the dict
and uses the try except catch brute force approach which was added
to workaround the issues with local testing. It's not pretty, but
it at least works.

Change-Id: Ic335ace8339b5f2ac07ba8fcf8ad125e1f816ae9