Improve labels on first and second pages
This commit replaces Build in any of the labels with Job on the first
and second pages. The term 'build' feels a bit misleading because in
the future we want to cover all job types. It also renames the labels
on the second page to be a bit more clear.

Change-Id: I25c157f2943994c3454825d188ed554afbed7c23