Add list of routes to / on api server
This commit adds a list of available routes as the response for a GET
on the root of the api server. It's a bit bizarre for end users to get
a 404 response when they do a GET on the API server. This will provide
users with a pointer to what's available.

Change-Id: I4da61ee74ec11bde96a99d1673e311155f126bc9