Fix run pass or fail evaluation
Previously, if a run doesn't have any test passing or failing, it was
accounted as a run that passed.

This was portraying an optimistic view of the health of the projects.

This commit shows that code for evaluating what is a passing run is
scattered across python and different javascript functions. This is
something that should be addressed on a subsequent commit.

Change-Id: Icda130607da6ef9835a918ffb2943f0585d6c049
Co-Authored-By: Moises Trovo <mtrovo@thoughtworks.com>