Remove '/test_runs' api endpoint
The api endpoint '/test_runs' is not being used by any page. As of
today, getting all this data from the database at once will simply end
up in a server error. Neither a browser nor the Apache server are ready
to handle this much data.
If we really want to have this, it will be better filtering it.

Change-Id: I3eee8b76e8a87c129592edce37cb199310a8cf1c