Fix ordering of project gauges on the home page
The project gauges used to be sorted by failure rate, but a recent
change broke this behavior (_.sortBy and Array.sort work in slightly
different ways).

This refactors the comparison function to comply with Array.sort
instead. It also changes the current test data so that we can test this
behavior as well.

Co-Authored-By: Dhiana Deva <ddeva@thoughtworks.com>
Co-Authored-By: Caio Carrara <ccarrara@thoughtworks.com>
Change-Id: Ifd9c379d805e2e62b3d89b54a6544974a8d6972a