Fix chart data loading
Recently we noticed that the charts on the first page were not being
loaded once the data was retrieved from the API. This was happening due
to controller variables being empty once it was instantiated.

Moving the assignment of these variables to happen inside the controller
method ensures the chart will only be loaded once they are available.

Change-Id: Ief661046d1647e1084c3b9f5eb43a6596e6f058e