Refactor home controller tests
The previous tests knew too much about the internals of HealthService
(i.e. they knew that it was using $http to get its data and even the
precise endpoint it was hitting). It also made it difficult to use
different data for different tests.

This refactors these tests so that we inject a mock for HealthService
and only stub the methods actually being used. This allows us to
completely ignore the way HealthService works internally while, at the
same time, allowing us to use different data on a case by case basis.

Change-Id: I192b973cfaa6c061a5f7e29f2625bcee177d22f1
Co-Authored-By: Caio Carrara <ccarrara@thoughtworks.com>