Refactor home controller process data test
This test used to test more functionality than it was meant to. In
particular, it should only be concerned about verifying whether the
passes/failures were correctly calculated, but instead, if we changed
the ordering of the projects, it would also fail.

This modifies it to be concerned only with the calculated project data.

Change-Id: Icfebbb8ef3765277ed3c0f21fe6f3ad53c2627cb
Co-Authored-By: Caio Carrara <ccarrara@thoughtworks.com>