Refactor project api and ui to make them generic
As part of providing a mechanism to let users group dashboard
information by other run metadata keys (project, voting, etc.), a few
changes were needed both on the ui and api.

On the api, the changes consist on making the endpoint accept any run
metadata key, and not only projects.
On the ui, the changes consist on renaming and adapting ProjectController
to become GroupController: a generic controller that is going to handle this
top-level grouping.

Change-Id: Ifde87132b66053b4a86b5d96c337a76e5554713e