Add factory to create Project models
This is the first step toward removing the business logic that currently
lives in the controllers, making them responsible for too much, into
specific objects that know how to perform them.

This allows us to have better tests and will also allow our controller
to be a lot thinner in the following commits (also allowing for
better/easier tests).

Co-Authored-By: Caio Carrara <ccarrara@thoughtworks.com>
Change-Id: I5499876b9fbbfcd616a2f5d823c416f4fee3ddbe