Make date window for job pages longer for periodic jobs
Periodic jobs by their nature run far less frequently than pre-merge
ci jobs. As such the graphs showing the per job statistics aren't really
meaningful when looking at a single day. This commit makes the window
as big as the other pages when being run for a periodic job. The window
is still kept small for other jobs because of performance reasons.

Eventually we want to have all the graphs have a date window slider to
make the generated results as big or as small as required. This will
likely be made obsolete when this is implemented. But, in the meantime
lets make things at least moderately useful.

Change-Id: Icb32a8784ef70b2b7ff2e95f7aa12371b7dbaac1