Rework get_test_runs_for_test() rest api
This commit reworks the get_test_runs_for() rest api call and changes
it's basic behavior. The first thing done is the numeric data is split
from the non-numeric data in the response. This was done to enable
mean based resampling on the numeric data. This is necessary because
the other large change to the api is that a running mean and a running
std dev are added to the response. This was done to enable more
detailed graphs to the UI for the run_time of a single test.
To perform the running mean and std dev pandas is added to the
requirements for the rest api.

Change-Id: I3416dc0cad24c90405ac437df31c39b59f524d88