Remove the remaining mockups
At this point the pages which we originally mocked up have all been
implemented. They served as a useful reference when we were first
bootstrapping the project, but we've grown beyond them now. This
also removes the last remaining bits of the license quagmire we
had with using highcharts since it is no longer in the code base
at all.

Change-Id: Icc3173b959ec0417f5ecd3804cf10d7d41ebfbe4