Fix mean runtime value on jobs page
This commit fixes the calculation of the mean runtime on the jobs page.
Previously it was just summing the runtime returned from the rest api,
which isn't really a mean. While taking a moving average of a bunch of
moving averages doesn't lead to the best value it is still better than
a sum.

Change-Id: I010b99c4a038ebd9154190ef606449b5cd3e2988