Properly interpolate missing data for test_runs
This commit modifies the data interpolation of the test_run_aggregator
method used by the per test view. Previously the commit just removed
rows post resample and d3 would do a linear interpolation between
points for missing data points. However, this is disingenous because
it gives the illusion that there were results for that period. This
commit switches to use pandas' index aware interpolation with a limit
of 20 (the same as our sample size for the rolling mean and std dev
calculations) consecutive interpolated points. This way we show large
gaps in the data properly, but smaller gaps are treated like they
were previously.

Change-Id: I8998faabe4e7fb523e2a8b5cdddb2b9e16e46e47