Cut the date window for test pagess in half
This commit changes the date window for the test pages in half from
30 days to 15 days. After fixing the data interpolation to properly
fill gaps this results in a more data being returned on the wire.
This can be a bit much for the browser to handle. So let's drop it
for now, the real fix here is to adding adjustable date windows and
selectable sample rates. But, those will take more time.

Change-Id: I4483dabbfbecff20eec57893ca14862d85b83191