Don't fix run_time to 2 decimal places
The per test pages were fixing the run_time value to 2 decimal places,
this was likely in an attempt to make the any written numbers be more
concise and readable. However when graphing the data, doing truncation
like that on the values makes things just look weird. Since we're not
actually printing any numbers we can just drop this. If we do need to
render text with the run_times in the future we can just do the
truncation in the html.

Change-Id: I467513cc62b6605ac3f8d1719bf042b133f14116