Set default type for stacked area to expanded
This commit changes the default graph type for stacked area to the
expanded type. The expanded stacked area graph type is by far generally
the most useful it clearly visualizes all the categories the best even
when one group is largely outnumbered. In the interactive view it also
shows a percentage which is normally more useful than a raw count. The
graph type is still user selectable for more specific use cases.

Change-Id: I4f446e3868990c748ecc85135031e2a84bcfa894