Properly show missing data in aggregated runs api response
This commit updates the runs aggregator to fill in the timestamps in
the response for missing data. Previously we were manually resampling
the data but only where there was valid data. This commit adds empty
lists to the response object where there is no data. This should
enable the js to properly represent the missing the data on the graph.

Change-Id: I95e37859783dc5597f8f0fc5e4f5e9ec1443d4f3