Properly display missing data in run graphs
This commit fixes the display of missing run data on the run stat
graphs. Previously d3 would interpolate between any 2 consecutive
data points regardless if there were missing samples between them.
In an earlier commit we updated the rest api to return an empty list
for samples that have no data. This commit builds off of that and
treats those empty lists as 0 for all values in that timestamp.

Change-Id: I137e4cfa1f22aaf2d932aa3d3420ebd841abca43