Make the tests page useful
This commit attempts to fix the tests page and make it useful. The
first thing it does is to remove all the bar graphs. These were added
to try and show something visual on the page. But they didn't convey
any information to the user and instead confused people.

The second thing this does is change the grouping to be more logical.
Instead of grouping on the test classes (which was also incorrect and
making assumptions about the path depth) this groups by the top level
path which will be the package the tests were in. (ie like tempest,
manila_tempest, etc) This makes more sense and also future proofs the
dashboard for when non-tempest tests are in the subunit2sql db.

The third thing this patch does is to change the default sorting order
on the tables to put the highest failure percentage at the top. This
will mean that what most people are interested in will be at the top
of the list. The fourth aspect of this patch is to default the groups
to open instead of closed.

Change-Id: I4426ba951276ef5d9482d67d5dfca7cd3d46d800