Handle uninitialized session in api methods
This commit adds a new method to get a session which attempts to
handle the case when the SessionMaker object wasn't initialized. We
were previously relying on flask to ensure things were properly
initiailzed but we've seen in production cases where Session was
None. This means that flask isn't initializing the SessionMaker
before the requests (or it's failing for some reason) This patch
will check if the session maker exists before it tries to call it
to get a session. If it doesn't setup is run to intialize everything
before we return the session object.

Change-Id: I257ead3d06bfc9832d3d109bd724e91277c3f175