Handle empty response in the run aggregator
This commit adds a check for the case when the run_aggregator is
passed an empty list from the db. We ere seeing cases on the
production server when an exception is being raised by trying to
run min() on an empty list. After this commit an empty response
will be returned before we attempt to do any additional processing
on the response.

Change-Id: I35dfc79da966650aa1bf80e46eeb68474fb3bf12