Improve pandas usage in test_runs_aggregator
This commit makes several improvements around the usage of pandas
in the test_run_aggregator. There were several issues in the usage
around edge conditions which weren't caught previously. First if
a results were passed in without any successes things would explode
when trying to resample the empty dataframe of successful runs. The
second failure was around the usage of interpolate. It was improperly
assigning the result of a dataframe interpolation to the run_time
series as part of the frame. How this worked is unclear, but we were
tripping issues on the production server from the incorrect assignment.
Finally the rolling mean and rollling std dev calculations only need
to be run post interpolation, so the earlier calls to these methods
are removed since they're not used at all.

Change-Id: I9e929c6cd1ffde9b389b6592b4f53fcccf711c30