Switch to a scatter plot for test runtime
This commit converts the runtime graph on the per test page to use
a scatter plot instead of a line chart. Since there are often large
discontinuities in the graph and the data is inherently noisy a line
plot can be problematic. Using a scatter plot will be more
representative of what we're trying to accomplish with the plot.
As part of this effort rows points with NaN for the value, mean, and
std dev are all dropped from the api response.

Change-Id: I87baf1ce98b4729702045a8e0046cbc49f372a0e