Fix possible race condition when loading recent runs.
The list of recent runs currently is not initialized to its default
value until processData() has completed. If the request for recent
runs happens to finish first, it can throw an error when attempting
to modify the array. This moves the array initialization to the
beginning of the function, and allocates it directly in the request
handler to prevent the error from occurring.

Change-Id: Id8aec75d2233617322f9ac24fdb1904f0b46f846