Fix invalid date NaN on 2nd graph on details page
This commit fixes the bug that is when rolling over the 2nd graph with
the mouse, some of the data is not displayed correctly, "NaN/NaN/0NaN"
is displayed instead.

When we use for-in loop against an object, we get a string instead of
integer. So in this patch, I just convert it with parseInt().

Closes-Bug: #1524088
Change-Id: I9d23e5be35c0fd2e0b19fd981db2c9fae476da54