Fix missing bare Session() call
This commit fixes a bare call to Session(). There was a previous
fix/workaround to the issue where the Sessionmaker factory Session()
isn't properly initialized when it was being called in an api method.
The get_session() method was added to reinitialize it if it's not
defined. However, 1 api method wasn't using this method which resulted
in a chance we'd still hit the other bug.

Change-Id: I6870ad9d0d1fe61f8783fbfb1e4e341cfb610058