Apply quotes rule for js files
This commit applies the quotes rule to keep the consistency. It seems we
and the Google style guide prefer single-quotes over double-quotes.
That's the reason why I choose single-quotes.

I don't want to stick this kind of things, actually. But I often confuse
that 'which quotes should I use?'. This commit avoids it.

Change-Id: Ic686d7e6bc91d8b75bba033c0b265b48a522df44