Improve projects list UI in the home page
This commit introduces bar graphs and a color classification by failure
rates to the home page project list.
In this commit, the width of the bar graphs is fixed to 300px because
auto widths cause unaligned widths. We should solve this issue in the
future but we don't have the solution right now.

Closes-Bug: #1543635
Change-Id: I448876d5342b1569ae7fe90c4a947dadab758a86