Split tests page into list and detail page
This commit splits the tests page into a list page and its detail page
for improving the performance of showing the page.
Previously, we rendered numerous tests list in one tests page. So it
causes the performance issue. Splitting the page reduces the number of
showing test lists. This is not perfect solution but it's better than
before at least.

Change-Id: If1a97740d09359ae17d0f0b7df8d36256ac99f47
Closes-Bug: #1548354