Improve status table CSS.
This adds several small improvements to the CSS for the status table
on the homepage. The table itself is now seamlessly enclosed in the
outer panel, and the bar chart column is now capped at 300px with
empty space to the side removed. Additionally, the chart is hidden
on small and extra-small layouts to prevent the table from overflowing
when there isn't enough space for it to fit.

Change-Id: Id17419e381d3a4e18caf10c4a2c96ad152c28963