Add warning for stale subunit2sql data
This commit adds a warning if the last gate run is > 1 day old in the
subunit2sql db. This is likely an indicator that something with
collection is broken and needs to be addressed. But, it can also be
confusing to users if there is no recent data, since the main pages
won't really show this well unless we miss a month's worth of data.

Change-Id: Ic1b9c4da371ca68919ac3f7fa6d2202dcc4f3720
Closes-Bug: #1550413