Add CSS rules to improve table column sizes
This adds a new CSS class, 'default-cols', based on the table CSS for
the homepage's project status table, but appropriate for most tables
on other pages. These rules allow word wrap for extremely long
columns in some cases, while cooperating with the 'table-responsive'
class. This replaces the 'status-table' class in places where it was
causing overflow with excessively large columns.

Change-Id: I7f88af7ed100085edeb57664177219e427df301a