Include build output in `npm run test` logs
This adds a gulp task that runs at the end of `gulp e2e` to copy
the built production site into the 'reports/' directory. This
directory will be uploaded to the log server during the npm test jobs
and should be viewable from browsers for quick evaluation of
client-side patches. A default config file is also included that
points to the official http://health.openstack.org/ API server so
this build preview can work against live data.

Change-Id: Ideb59ae050f77dab99a4f025e27519fd0fd11f6f