Update angular-ui-bootstrap
This updates angular-ui-bootstrap to the latest version. Among other
fixes, this corrects some odd display behavior when showing the period
selection popover, and allows us to enable dismiss-on-outside-click
functionality.

Note that as of 1.0.0, directives from angular-ui-bootstrap are
prefixed with 'uib-'. The dropdowns in the crumb menu have been
updated to account for this.

Change-Id: I04ff7d057ca881295c23b50eb277a60add384fbd