Adjust tooltip position for homepage gauges
This commit changes the tooltip placement for the horizontal gauge
charts on the homepage. Previously they would be rendered to the right
of the charts, but this could easily get cut off and cause a
horizontal scrollbar to appear. Tooltips are now placed above the
charts, and animations have also been disabled to avoid any lag
caused while scrolling or moving over many charts.

Change-Id: I6061fb634002ffefc1dea785621bd098bed32331