Change header of homepage status panel to camel case
This commit changes the header of the homepage status panel to camel
case. Originally, only first words was lowercases such as 'project
Status', 'filename Status', etc. It's a little wired. So This commit
fixes it. And this commit changes snake cases to camel cases, too. For
example, 'build_branch Status' -> 'Build Branch Status'.

Change-Id: I2a3cb01640c83d5211bad7831be8a3ff67049629