Remove 'Project' from status table column header
This commit removes the description 'Project' from the homepage status
table 2nd column. This table shows not only 'Project' status but also
build_branch, node_provider, filename... by grouping keys. So 'Project'
Name is not good at this column. And we already show the grouping key as
a table title.

Change-Id: I36daa85f77b2f84ac72d86fc370e09a3e1e37b1a