Use prefix API on tests and tests detail pages
This uses the new prefix APIs to provide server-side processing for
listing tests by prefix and pagination. This removes the need for
users to download the entire list of all tests (~4 MB of JSON taking
15 seconds to generate and download), instead sending only a list of
prefixes which the user can select from (~6 KB and < 0.5 seconds to
load). Additionally, the tests detail page now loads a paginated list
of tests by prefix from the server, greatly reducing the required time
and memory needed to load the page.

Change-Id: I103f77c63c4cb8200db842f48c514b4a214ef9b7