Add recent/detail openstack-health api.
Added a new api to retrieve the full details of the most recent runs, rather
than just the summary. This data was already available in the code, the api
simply returns it to the caller. This api can be used to provide “Test %”
complete rather than a simple “Runs passed/failed” which is in the current
/recent api.

Moved the existing functionality of the recent api call to a wrapper
module. This will be called by the recent api and the recent/detail api.
The detail version will return all the information returned by
subunit2sql and will not calculate pass or fail.

Added a unit test for the new recent/detail api.

Change-Id: I5528444148f525285cafed3fa9ce0a94dbfd5306