Handle urlencoded input parameters
This commit adds a call to urllib unquote() for every url parameter
we expect on each call. This is done as an explicit step instead of
trying to bake it into flask somehow to ensure we maintain backwards
compat with existing calls from the fronted, to avoid having to update
the frontend for this. This is not to maintain backwards compat in
general which is something we don't provide guarantees on.

Change-Id: I893375e1409dddfeb41f5a2ba8023cd3ddc5725a