Fix race condition in gulp prod and dev tasks
This fixes a race condition in gulp tasks where the 'browserify' task
would occasionally run before the 'views' task had finished, causing
an exception to be thrown when it failed to find the generated
templates. This makes the 'browserify' task wait for all other tasks
to complete before running, so the templates will always be generated
first.

Change-Id: I17dd4cd234dc61b98ba9f0e5ab527d8798e28434