Redesign failed tests panel
This redesigns the failed tests panel, replacing the panel/accordion
group/table combination with a flat table that should be much easier
to extend. This also adds a new 'nest' directive to better support
dynamic hiding/showing of content. The directive is able to insert
nested content as a sibling of any normal data row, allowing for
more consistent layouts.

Change-Id: Ic0ad498049ca0f008fbcecb1dbbac3c48ab9cb41