Fix crumb menu on test details page
Currently, the crumb menu's parent page detection is broken on the
tests detail page following the split of the original tests page into
'tests' and 'testsDetail', causing it to mistakenly show 'Overview'
instead of 'Tests'.  This updates the crumb menu's view code to
additionally check for the 'testsDetail' state.

A new filter, 'anyState', is also added to provide the necessary
functionality, since angular's own conditionals don't work in this
context.

Change-Id: I758f8230442607150c59be3b83ac7586679047ec