Skip ES lookups on runs with missing metadata
This commit fixes a bug when the data returned from subunit2sql is
missing the required fields, like in the case of periodic jobs, to
use the elastic-recheck classifier. Previously it was assumed all
runs being returned from subunit2sql would have an associated change
and the metadata associated with that, but this wasn't always true.
So now it will just skip trying to do an ES lookup if all the required
fields aren't present.

Change-Id: I395656c94499f658df1dd20cafa0783e394bc382