Update dogpile cache with an async worker
This commit changes how we refresh our cache from doing it directly on
stale requests to do it async in the background with a worker thread.
This leverages dogpile.cache's async_creation_runner mechanism to start
a background thread when the cache goes stale that will update the cache
while still returning the cached copy in them meantime.

To enable this you need to use a dogpile.cache backend that supports
using a distributed_lock. This does not include the default dbm backend,
using a memcached based backend, like dogpile.cache.pylibmc, is
recommended for enabling this functionality.

Change-Id: I0fd29839c72ca2fdfb4c4724bb3da7e283e3d27d