Add DBM cache backend that uses memcached locking
This commit adds a new dogpile.cache backend that uses the memcached
distributed locking mechanism to enable async workers but still uses
the dbm file storage. This should enable us to cache large JSON blobs
without worrying about the memacached size constraints, but at the
same time reap the benefits of having an async worker update the
cache in the background. The tradeoff here is configuration complexity
because you still need to install memcached to leverage this.

Change-Id: Ied241ca1762c62a047bd366d7bd105028a884f30