Remove failover logic outside of flask context
This commit removes some failover logic to make things in the get
recent test failure list api run without a flask context. This was
originally added in an earlier version of the async cache worker
implementation to handle directly calling the method outside of an
api request. (to refresh the cache) However, recent changes to that
have made this unecessary as the inner function is all that's called
by the async thread.

Change-Id: If0e7e621fec6c20f52a4dca162272b7296681ffc