Turn off interpolation on raw test run numeric data
This commit removes the interpolation stage from the test_run numeric
data processing. The interpolation is useful when we're dealing with
a line plot since it can fill the gaps for missing data. But, since we
moved to graphing the numeric data as a scatter plot the interpolation
does more harm then good. It basically forms a line artificially in
the data wherever there is missing data. This is both confusing and
misleading so let's just not do it. However, the interpolation is
still useful for generating the mean and std dev data. Those are
essentially line plots (although because of nvd3 limitations are
still displayed as scatter points) and having the gaps filled makes
it more useful, so we want to still use the interpolated data for
those values.

Change-Id: I55c40c6ca7145848c0da73c934851c20994df8d1